pyproject.toml
[tool.poetry]
name = "salvage"
version = "1.0.0"
description = "Backup Portainer Stacks to GitHub and report changes via Discord."
authors = ["EthanC <16727756+EthanC@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
loguru-discord = "^1.0.3"
httpx = "^0.24.1"
PyGithub = "^1.59.0"
discord-webhook = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
